+++ react/components/molecules/chat-message/chat-message.stories.js
@@ -1,4 +1,5 @@
 import React from "react";
+import { View } from "react-native";
 import { storiesOf } from "@storybook/react-native";
 import { withKnobs, number, text } from "@storybook/addon-knobs";
 
@@ -20,21 +21,30 @@ storiesOf("Molecules", module)
         const attachments = [
             {
                 name: "dummy.pdf",
-                url: "https://www.w3.org/WAI/ER/tests/xhtml/testfiles/resources/pdf/dummy.pdf"
+                path: "https://www.w3.org/WAI/ER/tests/xhtml/testfiles/resources/pdf/dummy.pdf"
             },
             {
                 name: "image.png",
                 path: "https://ripe-core-sbx.platforme.com/api/compose?brand=dummy&model=cube"
             }
         ];
+        const status = text("Status", "created");
 
         return (
-            <ChatMessage
-                avatarUrl={avatarUrl}
-                username={username}
-                message={message || undefined}
-                date={date || new Date()}
-                attachments={attachments}
-            />
+            <View>
+                <ChatMessage
+                    avatarUrl={avatarUrl}
+                    username={username}
+                    message={message || undefined}
+                    date={date || new Date()}
+                    attachments={attachments}
+                />
+                <ChatMessage
+                    avatarUrl={avatarUrl}
+                    username={username}
+                    status={status || undefined}
+                    date={date || new Date()}
+                />
+            </View>
         );
     });